backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
scikit-learn
pandas
numpy
joblib
python-dotenv
aiofiles
sqlite
groq
